Changed class layout
