+++ src/wemos.cpp
@@ -1,33 +1,8 @@
 #include "wemos.h"
 
-Stoel::Stoel(int fd, std::string clientName, std::string senderName){
-    this->clientName = clientName;
-    this->senderName = senderName;
-    this->fd = fd;
-    this->wemosType = "stoel";
-}
-
-Stoel::~Stoel(){
-}
 
 
 
-void Stoel::sensorUpdate(bool newStatus){
-    if(newStatus = status){ // aka status is the same, therefor useless
-        return;
-    }
-    if(newStatus == true){
-        turnOnTE();
-    }else{
-        turnOffTE();
-    }
-    return;
-}
 
-void Stoel::turnOnTE(){
 
-}
 
-void Stoel::turnOffTE(){
-
-}
