+++ src/wemos.h
@@ -5,27 +5,20 @@
 #include "xmlwriter.h"
 
 class Wemos {
+private:
+
 protected:
     int fd;
     std::string clientName;
     std::string senderName;
     std::string wemosType;
+    // TODO fd and type can be removed
 
 public:
 
     //Getters
-    virtual const bool getStatus(std::string sensor) = 0;
+    //virtual const bool getStatus(std::string sensor) = 0;
+    virtual std::string handleSensorUpdate();
+    virtual std::string handleSiteRequest();
 };
 
-class Stoel : private Wemos {
-private:
-    bool status;
-
-public:
-    Stoel(int fd, std::string clientName, std::string senderName);
-    ~Stoel();
-    inline const bool getStatus(std::string sensor)
-    void sensorUpdate(bool newStatus);
-    void turnOnTE();
-    void turnOffTE();
-};
